Reworked ArgumentRegistry Added TabCompletion
