+++ src/main/java/pl/plajer/pinata/commands/arguments/data/CommandDescription.java
@@ -1,0 +1,51 @@
+/*
+ * Pinata plugin - spawn pinata mob and kill it to get drops
+ * Copyright (C)2018 Plajer
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */package pl.plajer.pinata.commands.arguments.data;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+/**
+ * @author Plajer
+ * <p>
+ * Created at 25.11.2018
+ */
+@Retention(RetentionPolicy.RUNTIME)
+public @interface CommandDescription {
+
+  /**
+   * Main command description
+   *
+   * @return description
+   */
+  String description();
+
+  /**
+   * Descriptions for each permssion for command/argument
+   *
+   * @return list of descriptions
+   */
+  String[] permissionsDesc();
+
+  /**
+   * Additional command notes like deprecated usage, warnings etc.
+   *
+   * @return additional notes
+   */
+  String notes() default "";
+
+}
